requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.0=pypi_0
allennlp=2.8.0=pypi_0
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
autocommand=2.2.2=pypi_0
av=10.0.0=pypi_0
backcall=0.2.0=pypi_0
backports-csv=1.0.7=pypi_0
base58=2.1.1=pypi_0
beautifulsoup4=4.11.1=pypi_0
bleach=6.0.0=pypi_0
blessed=1.20.0=pypi_0
blis=0.7.9=pypi_0
boto3=1.26.79=pypi_0
botocore=1.29.79=pypi_0
ca-certificates=2022.07.19=h06a4308_0
cached-path=0.3.2=pypi_0
cachetools=4.2.4=pypi_0
catalogue=2.0.8=pypi_0
certifi=2022.9.24=py38h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pypi_0
checklist=0.0.11=pypi_0
cheroot=9.0.0=pypi_0
cherrypy=18.8.0=pypi_0
click=8.1.3=pypi_0
comm=0.1.3=pypi_0
contourpy=1.0.7=pypi_0
cryptography=40.0.1=pypi_0
cycler=0.11.0=pypi_0
cymem=2.0.7=pypi_0
dassl=0.6.3=dev_0
datasets=1.18.4=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.6=pypi_0
docker-pycreds=0.4.0=pypi_0
emoji=2.2.0=pypi_0
entrypoints=0.3=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.1.1=pypi_0
executing=1.2.0=pypi_0
fairscale=0.4.0=pypi_0
fastjsonschema=2.16.3=pypi_0
feedparser=6.0.10=pypi_0
filelock=3.3.2=pypi_0
flake8=3.7.9=pypi_0
fonttools=4.38.0=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.3.0=pypi_0
ftfy=6.1.1=pypi_0
future=0.18.2=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gdown=4.5.1=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-api-core=2.11.0=pypi_0
google-auth=2.17.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-cloud-core=2.3.2=pypi_0
google-cloud-storage=1.44.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-resumable-media=2.4.1=pypi_0
googleapis-common-protos=1.59.0=pypi_0
gpustat=1.1.1=pypi_0
grpcio=1.49.1=pypi_0
h5py=3.8.0=pypi_0
huggingface-hub=0.1.2=pypi_0
idna=3.4=pypi_0
imageio=2.22.4=pypi_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=5.12.0=pypi_0
inflect=6.0.1=pypi_0
iniconfig=2.0.0=pypi_0
install=1.3.5=pypi_0
iopath=0.1.10=pypi_0
ipykernel=6.22.0=pypi_0
ipython=8.11.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.6=pypi_0
iso-639=0.4.5=pypi_0
isoduration=20.11.0=pypi_0
isort=4.3.21=pypi_0
jaraco-collections=4.0.0=pypi_0
jaraco-context=4.3.0=pypi_0
jaraco-functools=3.6.0=pypi_0
jaraco-text=3.11.1=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.2.0=pypi_0
jsonnet=0.19.1=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.1.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.3.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.5.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.7=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
littleutils=0.2.2=pypi_0
lmdb=1.3.0=pypi_0
loguru=0.6.0=pypi_0
lxml=4.9.2=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.7.0=pypi_0
matplotlib-inline=0.1.6=pypi_0
maxflow=0.0.1=pypi_0
mccabe=0.6.1=pypi_0
mistune=2.0.5=pypi_0
more-itertools=9.1.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
munch=2.5.0=pypi_0
murmurhash=1.0.9=pypi_0
nbclassic=0.5.4=pypi_0
nbclient=0.7.2=pypi_0
nbconvert=7.2.10=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=pypi_0
networkx=2.8.8=pypi_0
nltk=3.8.1=pypi_0
notebook=6.5.3=pypi_0
notebook-shim=0.2.2=pypi_0
numpy=1.23.4=pypi_0
nvidia-ml-py=12.535.133=pypi_0
oauthlib=3.2.1=pypi_0
ogb=1.3.4=pypi_0
opencv-python=4.2.0.34=pypi_0
opencv-python-headless=4.6.0.66=pypi_0
openpyxl=3.1.1=pypi_0
openssl=1.1.1q=h7f8727e_0
outdated=0.2.1=pypi_0
overrides=3.1.0=pypi_0
packaging=21.3=pypi_0
pandas=1.5.0=pypi_0
pandocfilters=1.5.0=pypi_0
parameterized=0.8.1=pypi_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pathy=0.10.1=pypi_0
patternfork-nosql=3.6=pypi_0
pdfminer-six=20221105=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=pypi_0
pip=23.0.1=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.2.0=pypi_0
pluggy=1.0.0=pypi_0
portalocker=2.6.0=pypi_0
portend=3.1.0=pypi_0
preshed=3.0.8=pypi_0
prometheus-client=0.16.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=11.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.7=pypi_0
pycodestyle=2.5.0=pypi_0
pycparser=2.21=pypi_0
pydantic=1.8.2=pypi_0
pyflakes=2.1.1=pypi_0
pygments=2.14.0=pypi_0
pymaxflow=1.3.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=pypi_0
pytest=7.2.2=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
python-docx=0.8.11=pypi_0
python-json-logger=2.0.7=pypi_0
pytorchvideo=0.1.5=pypi_0
pytz=2022.4=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=pypi_0
qtconsole=5.4.1=pypi_0
qtpy=2.3.1=pypi_0
qudida=0.0.4=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.9.13=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.6.0=pypi_0
sacremoses=0.0.53=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.2=pypi_0
seaborn=0.12.2=pypi_0
send2trash=1.8.0=pypi_0
sentencepiece=0.1.97=pypi_0
sentry-sdk=1.18.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=67.6.0=pypi_0
sgmllib3k=1.0.0=pypi_0
shortuuid=1.0.11=pypi_0
simplecrf=0.2.1.1=pypi_0
six=1.16.0=pypi_0
sk-video=1.1.10=pypi_0
smart-open=6.3.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
spacy=3.1.7=pypi_0
spacy-legacy=3.0.12=pypi_0
sqlite=3.39.3=h5082296_0
sqlitedict=2.1.0=pypi_0
srsly=2.4.6=pypi_0
stack-data=0.6.2=pypi_0
stanza=1.5.0=pypi_0
tabulate=0.9.0=pypi_0
tb-nightly=2.11.0a20221013=pypi_0
tempora=5.2.1=pypi_0
tensorboard=2.4.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-logger=0.1.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.6=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.17.1=pypi_0
thinc=8.0.17=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.10.10=pypi_0
timm=0.6.12=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.10.3=pypi_0
tomli=2.0.1=pypi_0
torch=1.9.0+cu111=pypi_0
torchaudio=0.9.0=pypi_0
torchtext=0.5.0=pypi_0
torchvision=0.10.0+cu111=pypi_0
tornado=6.2=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.12.5=pypi_0
typer=0.4.2=pypi_0
typing-extensions=4.4.0=pypi_0
uri-template=1.2.0=pypi_0
urllib3=1.26.12=pypi_0
wandb=0.12.21=pypi_0
wasabi=0.10.1=pypi_0
wcwidth=0.2.5=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.40.0=pypi_0
widgetsnbextension=4.0.7=pypi_0
wilds=1.2.2=pypi_0
xxhash=3.2.0=pypi_0
xz=5.2.6=h5eee18b_0
yacs=0.1.8=pypi_0
yapf=0.29.0=pypi_0
yarl=1.8.2=pypi_0
zc-lockfile=3.0.post1=pypi_0
zipp=3.9.0=pypi_0
zlib=1.2.12=h5eee18b_3
